requirements.txt
# Chess RL with Adaptive Optimization
# Dependencies untuk training di Google Colab

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Chess Library
python-chess>=1.999

# Reinforcement Learning Environment
gymnasium>=0.29.0

# Numerical Computing
numpy>=1.24.0

# Visualization & Logging
tensorboard>=2.14.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Progress & Utilities
tqdm>=4.66.0
pyyaml>=6.0
colorama>=0.4.6

# Jupyter/Colab Support
ipywidgets>=8.1.0
IPython>=8.0.0

# Chess Visualization
cairosvg>=2.7.0
Pillow>=10.0.0

# Web Interface (untuk visual demo)
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Stockfish Python wrapper
stockfish>=3.28.0
